adds database seeding to admin component. reroutes app.vue call from api to backend
